Update TYPED_TEST_CASE to TYPED_TEST_SUITE
