+++ hipcub/include/hipcub/backend/rocprim/block/block_adjacent_difference.hpp
@@ -36,27 +36,6 @@
 
 BEGIN_HIPCUB_NAMESPACE
 
-namespace detail
-{
-  // Trait checks if FlagOp can be called with 3 arguments (a, b, b_index)
-  template<class T, class FlagOp, class = void>
-  struct WithBIndexArg
-      : std::false_type
-  { };
-
-  template<class T, class FlagOp>
-  struct WithBIndexArg<
-          T, FlagOp,
-          typename std::conditional<
-             true,
-             void,
-             decltype(std::declval<FlagOp>()(std::declval<T>(), std::declval<T>(), 0))
-          >::type
-      > : std::true_type
-  { };
-
-}
-
 template<
     typename T,
     int BLOCK_DIM_X,
